config: introduce sharding user role
This patch introduces the sharding user role. This role must be set to
the vshard storage user, if user is in 'credentials.user'.

Part of #8862

@TarantoolBot document
Title: config: credential `sharding` role

The credential `sharding` role is a new credential role. This is the
default credential role, but it is different from other default
credential roles because it is created by the config module. Other
default credential roles are already described in the bootstrap.snap
file.

This role has different privileges depending on the replicaset sharding
role. For replicasets with the sharding `storage` role, the credential
`sharding` role will have rights to execute necessary `vshard.storage.*`
functions and the credential `replicaset` role. If the replicaset does
not have the sharding `storage` role, the credential `sharding` role
does not have any privileges.

A sharding storage user must have the credential `sharding` role among
their credential roles, if the user is in `credentials.users`. If the
user is not in `credentials.users` we do not check its privileges.