config: drop alert about credentials and RO mode
Before this patch a replica adds an alert regarding inability to write
new credentials from config on startup. In most cases these new
credentials are applied on a master and the warning becomes obsolete.
The problem is that it is not removed at this point. A manual
`config:reload()` is needed to flush it.

It makes the warning more confusing than helpful. Let's remove it for
now.

We can return the alert back in a future, but we should do it together
with a logic that drops the alert if the needed data is received from a
master.

Part of #8862

NO_DOC=bugfix