config: start console before first box.cfg() call
This commit changes the order of configuration appliers and move
`console` before `box_cfg`. The `console` applier is adjusted to
correctly interpret the configured socket file path before the first
box.cfg().

The main reason to change the order is to allow a user to call
`box.ctl.make_bootstrap_leader()` for a replicaset that starts in the
`replication.boostrap_strategy: supervised` mode.

Part of #8862

NO_DOC=The documentation already assumes that a control command may be
       issued on the console. No changes are required.