config: rework replication.peers
After this patch, replication.peers will accept a list of instance names
instead of a list of URIs.

Closes #9541

@TarantoolBot document
Title: replication.peers config option

The replication.peers option accepts a list of instance names that will
be set as upstreams. If the option is not nil, it cannot be empty. For
an instance, the option can only contain the names of the instances of
the replicaset to which the instance belongs.

Currently, if the option does not contain all instances of a replicaset,
except maybe one instance, the instance `bootstrap_strategy` option must
be set to `config` and the instance replicaset `bootstrap_leader` option
must be one of the instances from the `replication.peers`.