test: make wal_off/alter robust against system tuple format count
We have to update the test every time we add a new system tuple format
because it checks the number of spaces that can be created before the
tuple format count limit is reached. Let's loosen the space count check
to avoid that.

We do this for the new memcs engine, which will be introduced to
Tarantool Enterprise Edition, because it will create a tuple format to
serve internal needs.

Needed for tarantool/tarantool-ee#617

NO_DOC=test fix
NO_CHANGELOG=test fix