raft: leader steps off on `ER_WAL_IO`
Prior to this patch, the raft leader continued to send heartbeats even
if it can't write anything, so it is seen by others as alive.
As a result, once the leader encounters a disk error, it can't write
anything, but the elections do not start. Now the leader resigns on the
first encounter with `ER_WAL_IO` write error.

Closes #9399

@TarantoolBot document
Title: raft: leader resigns on the first encounter with `ER_WAL_IO`

* Now the leader resigns on the first encounter with `ER_WAL_IO` write
  error (the leader broadcasts this).