config: fix leader failing to start when uuids are set
Currently a leader fails to start with `attempt to index a nil value`
error, when all UUIDs are specified in config. This is caused by the
fact, that we try to apply names to the instance, which have not joined
yet, _cluster record doesn't exist.

We must skip such names. Alert will be dropeed automatically as soon as
the instance is joined.

Closes #9572

NO_DOC=bugfix

Co-authored-by: Alexander Turenko <totktonada.ru@gmail.com>