ci: rm workarounds from `install-deps-osx` action
The issue [1] is closed, and it looks like the problem doesn't exist
now. So removing the respective workaround from the action.

Also, removing the workaround with upgrading packages as it is not
needed anymore because now `brew` just gives a warning instead of
failing when the package is already installed.

[1] https://github.com/gevent/gevent/issues/1721

NO_DOC=ci
NO_TEST=ci
NO_CHANGELOG=ci