index: introduce weak reference
An iterator doesn't pin the index it was created for. Instead, it holds
a 'weak' reference to it, which is invalidated if the index is dropped.
This is achieved by checking if the index can be found in the space
cache by id every time the reference is accessed. Let's factor out this
weak reference concept into a separate class so that we can reuse it
because we'll also need it to implement the index scanner API.

Needed for #9568

NO_DOC=refactoring
NO_TEST=refactoring
NO_CHANGELOG=refactoring