config: report all alerts of the same type
Before this patch `config:info().alerts` reports only last alert of a
certain type, while others are skipped. It is a regression from commit
fa97cc0c12e9 ("config: introduce droppable alerts").

The idea of the fix is to eliminate a key from all the alerts that are
not to be dropped later except on configuration reloading.

Fixes #9586

NO_DOC=bugfix