lyaml: fix alias serialization
The #8350 was introduced by the commit b42302f52ada ("lua-yaml: enable
aliasing for objects returned by __serialize") so the patch is
effectively reversed.

The idea is to call all object __serialize methods recursively before
finding references. In the process we need to keep an eye on reference
structure of course.

Closes #8350
Closes #8321
Closes #8310

NO_DOC=bugfix