config: call extras.post_apply() if reload failed
This patch changes when extras.post_apply() is executed. Previously, it
would only run if a _post_apply() was successful and all previous reload
steps completed properly. It will now be executed, even if some reload
steps failed.

Note that if an error appears during startup, Tarantool will stop and
extras.post_apply() will not be executed.

Also note that if the startup or restart is successful, there will be no
change in behavior due to this patch.

Needed for https://github.com/tarantool/tarantool-ee/issues/643

NO_DOC=will be added in EE
NO_CHANGELOG=will be added in EE