test: fix fix_lua_loader_error_handling on macOS
This patch is a follow-up for the commit
b99a35412d0c65dd84209872654fec9aaf2f1d02 ("lua: fix loader behaviour in
case of syntax error").

This test uses full matching of the error message and assumes that the
temporary directory is short enough, so the full patch is less than 60
symbols (size of `lua_Debug.short_src`). But for macOS, the temporary
directory has a longer path, so the error message is shrunk and isn't
equal to the expected string.

This patch replaces full matching with matching of the last part of the
error message.

Relates to tarantool/tarantool-ee#585

NO_DOC=test fix
NO_CHANGELOG=test fix