trigger: support arbitrary func_adapter in trigger module
The patch drops the invariant that all the triggers are Lua ones.
It is needed to introduce persistent triggers, which will be implemented
with another func_adapter inheritant.

Note that we need to invalidate handlers returned by pairs if they are not
Lua objects - they are not referenced by Lua and can be deleted when
they are not referenced by iterator.

Part of #8663

NO_TEST=see next commit
NO_CHANGELOG=no behavior changes
NO_DOC=no behavior changes