box: add watcher shutdown
We stop client fibers in the process of Tarantool shutdown in order to
be sure that subsystems that will be shutdown later are not being used.
See the commit bf620650c429 ("box: finish client fibers on shutdown").

But there is one more way for client code to be executed - in watcher
callback. So let's shutdown watcher too. After shutdown watcher API is
usable so we can next shutdown client fibers but notifications are
stopped.

Part of #8423

NO_CHANGELOG=internal
NO_DOC=internal