app: make snapshot on SIGUSR1 using gc subsystem
We create a snapshot on SIGUSR1 signal in a newly spawned system fiber.
It can interfere with Tarantool shutdown. In particular there is an
assertion on shutdown during such a snapshot because of cord making
snaphshot. Let's just trigger making snapshot in gc subsystem in it's
own worker fiber.

```
  #5  0x00007e7ec9a54d26 in __assert_fail (
      assertion=0x63ad06748400 "pm_atomic_load(&cord_count) == 0",
      file=0x63ad067478b8 "./src/lib/core/fiber.c", line=2290,
      function=0x63ad06748968 <__PRETTY_FUNCTION__.6> "fiber_free") at assert.c:101
  #6  0x000063ad061a6a91 in fiber_free ()
      at /home/shiny/dev/tarantool/src/lib/core/fiber.c:2290
  #7  0x000063ad05edc216 in tarantool_free ()
      at /home/shiny/dev/tarantool/src/main.cc:632
  #8  0x000063ad05edd144 in main (argc=1, argv=0x63ad079ca3b0)
```

Part of #8423

NO_CHANGELOG=internal
NO_DOC=internal