iproto: don't hang on uncancellable iproto request
Currently if there is uncancellable iproto request Tarantool shutdown
will hang. Let's instead give it some time and then panic.

On this way it is good to make iproto_drop_connections() fail on timeout.
It is used in `box.ctl.iproto_lockdown` which is better to fail on
timeout than to hang indefinitely too.

In Tarantool CI which is run with TEST_BUILD set, we set the timeout to
the infinity. This is on par with current fiber_shutdown() behaviour.
We will not change the latter for a while because there is already
several tests that count on that. Also it is currently easier to test
that there is no hang than to test exit status.

Part of #8423

NO_CHANGELOG=internal
NO_DOC=internal

(cherry picked from commit 22d507d5e9681eeb1cf6b3f4514227c0d5259727)