port: introduce get_c_entries virtual method
The method allows to represent port contents as a list of port_c_entry.
It is handy to easily iterate over the port contents since port_c_entry
is actually a list of variants of supported by Tarantool types. It is
needed to easily process values returned from triggers.

Returned entries mustn't be destroyed since they don't own any resources.
Their lifetime is similar to the MsgPack obtained by port_get_msgpack -
it may be returned directly from the port, or it can be allocated on
fiber()->gc, so the caller is responsible for cleaning the region up.

NO_CHANGELOG=internal
NO_DOC=internal