netbox: check if connection was closed after entering Lua VM from worker
When the worker fiber enters the Lua VM, garbage collection can be
triggered and the connection owning the worker could get closed.

To the best of our knowledge, there is no convenient way to get notified
about such an event, so the best we can do here is check the connection
state in the worker's main 'work' routine.

Closes #9621

NO_DOC=<bugfix>
NO_TEST=<the test relies on Lua garbage collection state, tested manually
         by Alexander Turenko <alexander.turenko@tarantool.org> and me on
         config-luatest/vshard_test.lua>