fiber: allow creating system fibers during shutdown
In the commit d40ce0fa028a ("core: disable fibers creation after
shutdown started") we disable creation of new fibers in the process of
shutdown. This may cause subsystem shutdown hanging.

The thing is we need fiber pool working during shutdown. For example
vinyl engine uses it through "tx" endpoint. Fibers in the pool are
finished after idle timeout. So we may have a situation when there is no
idle fibers in the pool and we cannot create a new one.

Part of #8423

NO_CHANGELOG=bugfix for unreleased version
NO_DOC=bugfix for unreleased version