config: allow to set non-existent credential role
Prior to this patch, privileges in the "credentials" section of the
configuration could be requested for non-existent spaces, functions, and
sequences. However, this was not possible for roles, although
essentially the same mechanism is used. The problem is resolved with
this patch.

Part of #9643

@TarantoolBot document
Title: Assign a non-existent credential role

A non-existent credential role can be assigned to a user or role in
`credential` section of the config. The actual assignment will occur
when the assigned credential role is created.