config: add ordered dictionary (internal module)
In development of the configuration module we met several cases, when an
order of appearance of some items is as important as a constant time
access using a key.

This commit adds a collection that serves such a need. It is to be used
internally in the src/box/lua/config code.

The module is inspired by Python's collections.OrderedDict. See the
description in the module code for details.

NO_DOC=the module is for internal use, at least for now
NO_CHANGELOG=see NO_DOC