port_c: drop assertion in port_c_dump_lua method
The method actually supports all modes of dump to Lua, but the assertion
checks that mp_object mode is never used - the commit drops the
assertion.

This mode is tested with a tuple constraint taking raw args. Along the
way, some test cases are enabled back because the bug that caused them
to be disabled has been fixed.

NO_CHANGELOG=internal
NO_DOC=internal