config: move creation of role sharding
Move the creation of the credential sharding role to the apply() method
in credentials.lua. This allows us to reduce the dependency on config,
and we can later move the creation of the credential sharding role to
sharding.lua.

Part of #9643

NO_DOC=refactoring
NO_TEST=refactoring
NO_CHANGELOG=refactoring