bps: add optional child cardinalities to inner blocks
An inner block cardinality is amount of elements in all the leaf nodes
placed under it (through inner children if any, recursively).

Now it's possible to store the cardinalities in inner nodes. This will
allow to implement getting the tree element by offset with logarithmic
complexity.

The default BPS tree test is split into the generic and specific parts.
The generic part is reused by the new test.

Since creation of new tests with .result files is prohibited by the
checkpatch, the generic part had been modified to make the new test
TAP-compatible.

Also, given the new test's output is not checked, a bunch of new
assertions had been added to it to make it functional. And the
printing test had been disabled for it since it requires the
output check.

Part of #8204

NO_DOC=internal
NO_CHANGELOG=internal